pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "zzzzz167", email = "2742400566@qq.com"},
]
dependencies = [
    "graia-ariadne[full]>=0.11.5",
    "graiax-text2img-playwright>=0.4.0",
    "graiax-fastapi>=0.2.1",
    "kayaku>=0.5.2",
    "graiax-shortcut>=0.3.0",
    "arclet-alconna-graia>=0.14.0",
    "arclet-alconna-ariadne>=0.14.0",
    "SQLAlchemy>=2.0.15",
    "aiosqlite>=0.19.0",
    "psutil>=5.9.5",
    "aiofile>=3.8.5",
    "Pillow>=9.5.0",
    "fontTools>=4.39.4",
]
requires-python = ">=3.9"
license = {text = "MIT"}

